pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "keras>=2.14.0",
    "numpy>=2.2.5",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "tensorflow-cpu>=2.14.0",
    "tensorflow>=2.14.0",
    "werkzeug>=3.1.3",
]
